# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WhoEarns API
  version: '0.4.0'
  description: |
    Read-only HTTP API for WhoEarns, an AI-assisted open-source Solana
    validator income intelligence service by 0base.vc. Income is derived
    from Solana RPC block data only: base fees, priority fees, and on-chain
    Jito tips found in produced blocks.
    Numeric fields are `null` when the matching `hasSlots` or `hasIncome`
    boolean is false.

    **v0.3 breaking change.** Jito Kobe payout fields and per-stream status
    enums were removed. Current-epoch values are lower bounds and can increase
    until the epoch closes; closed-epoch rows are marked with `isFinal=true`.
    Repository migration note: `docs/migrations/v0.3-api.md`.

    **v0.4 leaderboard/search update.** `/v1/leaderboard` now defaults to
    `window=live_trend&sort=income_per_slot` and also supports
    `window=decade_epoch` for the latest complete 10-epoch ranking block.
    Validator name search is exposed at `/v1/validators/search`.
    Repository migration note:
    `docs/migrations/v0.4-api.md`.
servers:
  - url: http://localhost:8080
    description: Placeholder — runtime route rewrites `servers[0].url` to match `SITE_URL` at request time
paths:
...
  /v1/claims/{vote}/wallets:
    post:
      summary: Register an operator-controlled wallet against a claimed validator
      description: |
        Phase 3 Claim v2 — the operator co-signs a canonical nonce
        with BOTH the validator identity key AND the wallet key, and
        anchors the registration with a 0-lamport memo transaction
        whose signature is supplied as `anchorTxSignature`. The
        server verifies both Ed25519 signatures against the nonce
        bytes (which carry both pubkeys, preventing rebinding) and
        records the (vote, wallet) tuple with a 90-day expiry.
        Appends to a bounded collection (≤3 wallets), hence `POST` +
        the plural `wallets` path.

        The vote pubkey travels in BOTH the path and the request body;
        the body stays authoritative and a `params.vote` ≠
        `votePubkey` mismatch is rejected `400 vote_pubkey_mismatch`.
        Per-validator cap: **3 wallets** (cold / warm / multisig).
        Re-attestation is required every 90 days; lapsed wallets stop
        contributing scoring signal until re-registered. Self-
        registration (wallet pubkey equal to vote/identity) is
        rejected with `400 pubkey_role_collision` since it pollutes
        downstream activity analytics.
      parameters:
        - $ref: '#/components/parameters/VotePath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClaimWalletVerifyRequest' }
      responses:
        '200':
          description: Wallet registration recorded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClaimWalletVerifyResponse' }
        '400':
          description: Malformed body, `pubkey_role_collision`, or `vote_pubkey_mismatch` (path ≠ body).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: |
            Rejected — typically one of:
            `not_claimed`, `identity_mismatch`, `stale_timestamp`,
            `signature_invalid`, `anchor_tx_invalid`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: |
            `wallet_cap_reached` — validator already has the maximum
            number of registered wallets, OR a same-nonce replay was
            caught at the DB UNIQUE constraint.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: |
            Feature disabled (Phase 3 dependencies not configured).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/operator-wallets/{wallet}:
    get:
      summary: Registration metadata for a registered operator wallet
      description: |
        Parent resource for the `/activity` sub-path. Returns the
        wallet's registration metadata — which validator it belongs
        to, the operator-chosen label, and the registration / expiry
        timestamps.

        **Existence-oracle defense.** Gated identically to
        `/activity`: only ACTIVE (not-expired) registered wallets are
        exposed. An unregistered or expired-registration pubkey
        returns 404, so the route cannot be used to enumerate the
        registered-wallet set.

        The forensic `signedNonce` / `anchorTxSignature` columns are
        deliberately omitted — everything returned is already
        operator-published or derivable from the on-chain
        registration.
      parameters:
        - $ref: '#/components/parameters/WalletPath'
      responses:
        '200':
          description: Wallet registration metadata.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorWalletResponse' }
        '400':
          description: Invalid wallet pubkey.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Wallet is not a currently-registered operator wallet.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/operator-wallets/{wallet}/activity:
    get:
      summary: Daily on-chain activity for a registered operator wallet
      description: |
        Phase 4 wallet heatmap source. Returns up to `days` (1-365)
        of UTC-bucketed daily transaction counts for the wallet. Days
        with zero activity are **omitted** from the response — clients
        zero-fill at draw time, keeping the payload tight for sparse
        wallets.

        **Existence-oracle defense.** Only ACTIVE (not-expired)
        registered wallets are exposed. An unregistered or expired
        pubkey returns 404 to prevent enumeration of the registered
        wallet set.

        `txFeesLamports` is `null` in the Phase 4 release — the
        ingester ships tx counts only. The field is reserved for a
        future backfill that adds per-day fee totals. `null` is
        intentional (not `"0"`) so a consumer summing fees can detect
        the unavailable-data case.
      parameters:
        - $ref: '#/components/parameters/WalletPath'
        - name: days
          in: query
          required: false
          description: Number of trailing days to return. Clamped to [1, 365]. Default 365.
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 365
      responses:
        '200':
          description: Daily activity entries (newest first).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletActivityResponse' }
        '400':
          description: Invalid wallet pubkey or query parameter.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Wallet is not a currently-registered operator wallet.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  parameters:
...
    VotePath:
      name: vote
      in: path
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 44
        pattern: '^[1-9A-HJ-NP-Za-km-z]+$'
    WalletPath:
      name: wallet
      in: path
      required: true
      description: Solana base58 wallet pubkey.
      schema:
        type: string
        minLength: 32
        maxLength: 44
        pattern: '^[1-9A-HJ-NP-Za-km-z]+$'
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, description: Machine-readable stable identifier. }
            message: { type: string }
            requestId: { type: string, format: uuid }
            details: { type: object, additionalProperties: true }
...
    ClaimWalletVerifyRequest:
      type: object
      required:
        - votePubkey
        - identityPubkey
        - walletPubkey
        - timestampMs
        - identitySignatureB58
        - walletSignatureB58
        - anchorTxSignature
      properties:
        votePubkey:
          type: string
          minLength: 32
          maxLength: 44
          pattern: '^[1-9A-HJ-NP-Za-km-z]+$'
        identityPubkey:
          type: string
          minLength: 32
          maxLength: 44
          pattern: '^[1-9A-HJ-NP-Za-km-z]+$'
        walletPubkey:
          type: string
          minLength: 32
          maxLength: 44
          pattern: '^[1-9A-HJ-NP-Za-km-z]+$'
        label:
          type: string
          maxLength: 32
          default: ''
          description: |
            Operator-chosen short label (cold / warm / multisig / ...).

            Rejected characters (400 `validation_error`): angle
            brackets (`<`, `>`), backticks, braces (`{`, `}`), and
            Unicode bidirectional-control codepoints (U+200E,
            U+200F, U+202A-U+202E, U+2066-U+2069). The label
            renders into a public hub surface, so these are
            blocked to prevent stored-XSS / RTL-override
            phishing of surrounding chrome.
        timestampMs: { type: integer, minimum: 1 }
        identitySignatureB58:
          type: string
          minLength: 64
          maxLength: 128
          description: Base58 Ed25519 signature by the validator identity key over the canonical nonce.
        walletSignatureB58:
          type: string
          minLength: 64
          maxLength: 128
          description: Base58 Ed25519 signature by the wallet key over the same canonical nonce.
        anchorTxSignature:
          type: string
          minLength: 64
          maxLength: 96
          description: |
            Solana transaction signature for the operationally-alive
            memo (0-lamport self-transfer with a memo containing the
            nonce hash).
    ClaimWalletVerifyResponse:
      type: object
      required: [wallet]
      properties:
        wallet:
          type: object
          required: [walletPubkey, label, registeredAt, expiresAt]
          properties:
            walletPubkey: { type: string }
            label: { type: string }
            registeredAt: { type: string, format: date-time }
            expiresAt:
              type: string
              format: date-time
              description: 90 days from registration. Lapsed wallets drop out of activity scoring.

    # --- Phase 4 Wallet activity ------------------------------------------
    OperatorWalletResponse:
      type: object
      required: [wallet, vote, label, registeredAt, expiresAt]
      description: |
        Registration metadata for one ACTIVE operator wallet. The
        forensic `signedNonce` / `anchorTxSignature` columns are NOT
        included.
      properties:
        wallet:
          type: string
          description: The operator wallet's base58 pubkey.
        vote:
          type: string
          description: Vote pubkey of the validator this wallet is registered under.
        label:
          type: string
          description: Operator-chosen label (≤32 chars), e.g. "hot" / "cold".
        registeredAt:
          type: string
          format: date-time
          description: When the wallet registration was verified.
        expiresAt:
          type: string
          format: date-time
          description: |
            When the registration attestation lapses. Past this the
            wallet drops out of scoring and this endpoint 404s.
    WalletActivityResponse:
      type: object
      required: [wallet, days, entries]
      properties:
        wallet: { type: string }
        days:
          type: integer
          minimum: 1
          maximum: 365
          description: Number of trailing days requested (after server-side clamp).
        entries:
          type: array
          description: |
            Newest-first daily activity rows. Days with zero activity
            are OMITTED — clients zero-fill at draw time.
          items:
            type: object
            required: [date, txCount, txFeesLamports]
            properties:
              date:
                type: string
                pattern: '^\d{4}-\d{2}-\d{2}$'
                description: UTC date (YYYY-MM-DD).
              txCount: { type: integer, minimum: 0 }
              txFeesLamports:
                type: [string, 'null']
                description: |
                  Per-day fee total in lamports as a decimal string.
                  `null` in the Phase 4 release — the ingester ships
                  tx counts only. Reserved for a future backfill.

    # --- Phase 5 SIMD curation feed ---------------------------------------
...